main

+++ base/pipelines/inference.py
@@ -686,7 +686,7 @@ class VideoGenPipeline(DiffusionPipeline):
                     if callback is not None and i % callback_steps == 0:
                         callback(i, t, latents)
 
-        
+            latents = latents.to(self.unet.dtype)
             # 8. Post-processing
             video = self.decode_latents(latents)
